refactored and normalized database to build the following logic:
We have really four tables at play here. when i refer to a "unique member" i mean that within an owning collection, each child may appear once, but each child may appear in any number of collections.
GroupSets - A collection holding an arbitrary number of hierarchically organized groups - each group has a parentID and an ordinal to track its location within the groupset Groups -- which are unique members of GroupSets , these contain an arbitrary number of plugins, which also have an ordinal, enforcing the remainder of the hierarchy Plugins -- these are unique members of Groups within the context of a groupset, these are also unique within relation to themselves, only one version of the named plugin exists Profiles -- within the context of a LoadOutProfile, which is also correlated with a groupset, but it's a many to many relationship, The loadout profile defines for a given groupset which plugins are enabled or not (this is different than the groups which enforce ordinal relationships, as well as define presence in the output file) -- a plugin may exist in the output file or not, but if it does it must have an enablement status that the ProfilePlugins table tracks